Updated merge_players and some comments
The merge_players function is meant to edit mistakes when a player is entered
with two different names in the system, possibly due to a typo.

Some comments were added or updated. Lots of changes are still necessary, e.g.
will Indigo Plateau use some sort of automatic documentation mechanism? (It
should)